dev-requirements.txt
# dev
ipython
django-extensions

-r requirements.txt